Move formatQuery tests to dedicated tests folder
