+++ packages/react-querybuilder/src/utils/formatQuery/__tests__/dbquery.cel.test.ts
@@ -1,7 +1,7 @@
-import { verifyCELEvaluator } from '../../../../../utils/cel-evaluator/verifyCELEvaluator';
-import { transformQuery } from '../transformQuery';
-import { dbTests, superUsers } from './dbqueryTestUtils';
-import { formatQuery } from './formatQuery';
+import { verifyCELEvaluator } from '../../../../../../utils/cel-evaluator/verifyCELEvaluator';
+import { transformQuery } from '../../transformQuery';
+import { dbTests, superUsers } from '../dbqueryTestUtils';
+import { formatQuery } from '../formatQuery';
 
 const celEvaluator = await verifyCELEvaluator();
