+++ packages/react-querybuilder/src/utils/formatQuery/__tests__/dbquery.postgres.drizzle.test.ts
@@ -4,9 +4,9 @@ import { PGlite } from '@electric-sql/pglite';
 import { boolean, integer, pgTable, text } from 'drizzle-orm/pg-core';
 import { drizzle } from 'drizzle-orm/pglite';
 import { convertToIC, formatQuery } from 'react-querybuilder';
-import type { TestSQLParams } from './dbqueryTestUtils';
-import { dbSetup, superUsers } from './dbqueryTestUtils';
-import { dbTestsDrizzle } from './dbqueryDrizzleTestUtils';
+import type { TestSQLParams } from '../dbqueryTestUtils';
+import { dbSetup, superUsers } from '../dbqueryTestUtils';
+import { dbTestsDrizzle } from '../dbqueryDrizzleTestUtils';
 
 const columnsPostgres = {
   firstName: text().notNull(),
