+++ packages/react-querybuilder/src/utils/formatQuery/__tests__/formatQuery.CEL.test.ts
@@ -1,7 +1,11 @@
-import type { FormatQueryOptions, RuleGroupType, RuleProcessor } from '../../types/index.noReact';
-import { add } from '../queryTools';
-import { defaultRuleProcessorCEL } from './defaultRuleProcessorCEL';
-import { formatQuery } from './formatQuery';
+import type {
+  FormatQueryOptions,
+  RuleGroupType,
+  RuleProcessor,
+} from '../../../types/index.noReact';
+import { add } from '../../queryTools';
+import { defaultRuleProcessorCEL } from '../defaultRuleProcessorCEL';
+import { formatQuery } from '../formatQuery';
 import {
   getValidationTestData,
   query,
@@ -11,8 +15,8 @@ import {
   queryIC,
   queryWithValueSourceField,
   testQueryDQ,
-} from './formatQueryTestUtils';
-import { defaultCELValueProcessor } from './index';
+} from '../formatQueryTestUtils';
+import { defaultCELValueProcessor } from '../index';
 
 const celString =
   'firstName == null && lastName != null && firstName in ["Test", "This"] && !(lastName in ["Test", "This"]) && firstName in [] && (firstName >= "Test" && firstName <= "This") && (firstName >= "Test" && firstName <= "This") && (lastName < "Test" || lastName > "This") && (age >= 12 && age <= 14) && age == "26" && isMusician == true && isLucky == false && !(gender == "M" || job != "Programmer" || email.contains("@")) && (!lastName.contains("ab") || job.startsWith("Prog") || email.endsWith("com") || !job.startsWith("Man") || !email.endsWith("fr"))';
